move compile function to compiler file so that it is no imported by default
