+++ lib/compiler.js
@@ -93,3 +93,17 @@ export default class Compiler {
     }
   }
 }
+
+export async function compile(configPath) {
+  try {
+    let config = {};
+    if (configPath) {
+      const contents = fs.readFileSync(configPath);
+      config = JSON.parse(contents);
+    }
+    const compiler = new Compiler(config);
+    await compiler.compile();
+  } catch (err) {
+    console.error(err);
+  }
+}
