+++ lib/waffle.js
@@ -1,6 +1,4 @@
-import fs from 'fs';
 import Ganache from 'ganache-core';
-import Compiler from './compiler';
 import {ContractFactory, providers, Contract, Wallet} from 'ethers';
 import matchers from './matchers';
 import defaultAccounts from './config/defaultAccounts';
@@ -29,20 +27,6 @@ export async function deployContract(wallet, contractJSON, args = [], overrideOp
   const tx = await wallet.sendTransaction(deployTransaction);
   const receipt = await wallet.provider.getTransactionReceipt(tx.hash);
   return new Contract(receipt.contractAddress, abi, wallet);
-}
-
-export async function compile(configPath) {
-  try {
-    let config = {};
-    if (configPath) {
-      const contents = fs.readFileSync(configPath);
-      config = JSON.parse(contents);
-    }
-    const compiler = new Compiler(config);
-    await compiler.compile();
-  } catch (err) {
-    console.error(err);
-  }
 }
 
 export const contractWithWallet = (contract, wallet) =>
